+++ B0.ml
@@ -52,20 +52,17 @@ let blueprint_cmds = test ~/"test/blueprint_cmds.ml" ~run:false
 (* Completion scripts update *)
 
 let update_completion_scripts =
-  B0_unit.of_action "update-completion-scripts" @@ fun env _ ~args:_ ->
-  let zsh = B0_env.in_scope_dir env ~/"src/tool/zsh-completion.sh" in
+  B0_unit.of_action "update-cmdliner-data" @@ fun env _ ~args:_ ->
   let bash = B0_env.in_scope_dir env ~/"src/tool/bash-completion.sh" in
-  let ml = B0_env.in_scope_dir env ~/"src/tool/cmdliner_completion.ml" in
-  let* zsh = Os.File.read zsh in
+  let zsh = B0_env.in_scope_dir env ~/"src/tool/zsh-completion.sh" in
+  let ml = B0_env.in_scope_dir env ~/"src/tool/cmdliner_data.ml" in
   let* bash = Os.File.read bash in
-  let rename s = String.replace_all ~sub:"_cmdliner_generic" ~by:"_%s" s in
+  let* zsh = Os.File.read zsh in
   let src = Fmt.str
-      "let zsh_completion name = Printf.sprintf {|%s\
-       compdef _%%s %%s\n\
-       |} name name name\n\n\
-       let bash_completion name = Printf.sprintf {|%s\
-       complete -F _%%s %%s\n\
-       |} name name name" (rename zsh) (rename bash)
+      "let bash_generic_completion =\n{|%s\
+       |}\n\n\
+       let zsh_generic_completion =\n{|%s\
+       |}" bash zsh
   in
   Os.File.write ~force:true ~make_path:false ml src